+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 
 - Fixes to CLI
 - Handle invalid event-XML containing invalid trailing characters
+- Improve constructing metadata when playing media on DLNA/DMR devices
 
 
 0.14.5 (2019-03-02)